Rework FactorialCalculator to be thread-safe
Turns out there were race conditions happening in the old factorial
calculator, because several threads saw that the current value isn't yet
calculated and both pushed the new calculated value to the cache. But we
can just calculate all the values when the factorial calculator is
created and then just get them.